Small fix to sync command
When updating a local file, the modify date in the index needs to
be set to the modify date from the local file, whereas before I
was setting it to the remote value. That was stupid because it
meant as soon as the local file was created/update on the next
sync it would appear as if these files had been modified and it
would copy back the files again to remote.