Write file before index for Sneeze. Also use local mod dates
The Sneeze function wrote the index before writing the file, but
this makes no sense as means that can't used the local system
modifydate, have to use what Simplenote returns. Which is stupid
because when you run a sync it will immediately want to update
the remote with local changes (that didn't actually happen).

And on that note, for all functions we want to be making sure that
when the index gets updated it is the local system modifydate
that is being used as otherwise it will want to synchronise
files (even when there have beenno changes).))